requirements.txt
requests==2.26.0
fastapi==0.81.0
SQLAlchemy==2.0.15
uvicorn==0.22.0
starlette==0.19.1
pydantic==1.10.7
setuptools==65.5.1
python-dotenv